query_sql/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
chromadb==0.5.7
langchain==0.3.3
pymupdf==1.24.10
pandas==2.2.3
openpyxl==3.1.5
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-multipart==0.0.12
pgvector==0.3.0
httpx==0.27.0
tenacity==8.5.0
streamlit==1.36.0
psycopg2-binary==2.9.9
google-generativeai==0.8.3